+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.1
+
+* Make GrassSurface properties public
+* Rename SurfaceCondition::Dry to Inconspicuous as it can be combined with wet gras surface
+
 ## 0.2.0
 
 * Add social files ([README.md](README.md), [CONTRIBUTING.md](CONTRIBUTING.md))